added all

+++ UserService/src/main/java/com/manikanta/microservices/project/UserService/Service/Implementation/UserServiceImplementation.java
@@ -1,17 +1,19 @@
 package com.manikanta.microservices.project.UserService.Service.Implementation;
 
-//import com.manikanta.microservices.project.UserService.Controller.JWTService;
+//import com.manikanta.microservices.project.UserService.Security.JWTService;
+
 import com.manikanta.microservices.project.UserService.DTO.*;
 import com.manikanta.microservices.project.UserService.Entity.User;
-import com.manikanta.microservices.project.UserService.Enums.MyFeatures;
 import com.manikanta.microservices.project.UserService.Exception.EmailAlreadyFoundException;
 import com.manikanta.microservices.project.UserService.Exception.UserNotFoundException;
+import com.manikanta.microservices.project.UserService.FeignClients.FeignAPIClient;
+import com.manikanta.microservices.project.UserService.FeignClients.FeignNotificationClient;
+import com.manikanta.microservices.project.UserService.FeignClients.FeignWeblabClient;
 import com.manikanta.microservices.project.UserService.Mapper.AutoUserMapper;
+import com.manikanta.microservices.project.UserService.Publisher.NotificationPublisher;
 import com.manikanta.microservices.project.UserService.Repository.UserRepository;
-import com.manikanta.microservices.project.UserService.Service.FeignAPIClient;
-import com.manikanta.microservices.project.UserService.Service.FeignWeblabClient;
+import com.manikanta.microservices.project.UserService.Security.JWTService;
 import com.manikanta.microservices.project.UserService.Service.UserService;
-import io.github.resilience4j.retry.annotation.Retry;
 import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
 import org.slf4j.Logger;
@@ -19,21 +21,23 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.CachePut;
-import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
-//import org.springframework.security.authentication.AuthenticationManager;
-//import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-//import org.springframework.security.core.Authentication;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
-import org.springframework.web.reactive.function.client.WebClient;
+import org.springframework.web.server.ResponseStatusException;
 import org.togglz.core.manager.FeatureManager;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -41,11 +45,12 @@ import java.util.stream.Collectors;
 @NoArgsConstructor
 public class UserServiceImplementation implements UserService {
 
-//    @Autowired
-//    private JWTService jwtService;
+    @Autowired
+    private JWTService jwtService;
 
     @Autowired
     private FeatureManager featureManager;
+
 
     @Autowired
     private UserRepository userRepository;
@@ -66,19 +71,36 @@ public class UserServiceImplementation implements UserService {
     @Autowired
     private FeignWeblabClient feignWeblabClient;
 
-//    @Autowired
-//    AuthenticationManager authManager;
+    @Autowired
+    private FeignNotificationClient feignNotificationClient;
 
+    @Autowired
+    AuthenticationManager authManager;
+
+    @Autowired
+    NotificationPublisher notificationPublisher;
+
+    public void sendNotification(){
+        logger.info("Inside sendNotification method");
+        Notification notification = Notification.builder()
+                        .recipient("tsupriya1925@gmail.com")
+                        .subject("User Creation")
+                        .message("Account Created Successfully")
+                        .build();
+//        boolean notificationStatus = feignNotificationClient.sendNotification(notification);
+        notificationPublisher.publishNotificationCreatedEvent(notification);
+    }
 //    @Autowired
 //    private ModelMapper mapper;
 
 //    @Cacheable(value = "getUserById", key = "#userId")
     @Override
-    public UserDTO getUser(Long userId){
-        logger.info("entered into get user by id method");
-        User user = userRepository.findById(userId).orElseThrow(()->new UserNotFoundException("No User Found with this userId: "+userId));
-        System.out.println("Using the new checkout feature.");
-        return AutoUserMapper.MAPPER.mapToDTO(user);
+    public ResponseEntity<UserDTO> getUser(Long userId){
+        logger.info("inside getUser method");
+//        User user = userRepository.findById(userId).orElseThrow(()->new UserNotFoundException("No User Found with this userId: "+userId));
+        User user = userRepository.findById(userId).orElseThrow(()-> new UserNotFoundException("User not found"));
+        return new ResponseEntity<>(AutoUserMapper.MAPPER.mapToDTO(user), HttpStatus.OK);
+
 //        if (featureManager.isActive(MyFeatures.NEW_CHECKOUT_FEATURE)) {
 //            // New checkout feature logic
 //            User user = userRepository.findById(userId).orElseThrow(()->new UserNotFoundException("No User Found with this userId: "+userId));
@@ -94,34 +116,48 @@ public class UserServiceImplementation implements UserService {
 
     @Override
     public UserDTO getUserDefault(){
-        logger.info("entered into get user by id method");
+        logger.info("inside getUserDefault method");
         User user = userRepository.findById(1L).orElseThrow(()->new UserNotFoundException("No User Found with this userId: Default"));
-        System.out.println("Using the new checkout feature.");
         return AutoUserMapper.MAPPER.mapToDTO(user);
     }
 
 
     @Override
     @CacheEvict(value = "getUserById", key = "#userId")
-    public void deleteUser(Long userId) {
+    public ResponseEntity<String> deleteUser(Long userId) {
+        logger.info("inside deleteUser method");
+        User user = userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException("User Not Found"));
+        if(user == null){
+            return new ResponseEntity<>("User Not Found",HttpStatus.NOT_FOUND);
+        }
         userRepository.deleteById(userId);
-        System.out.println("Deleted");
+
+        return new ResponseEntity<>("Deleted "+userId,HttpStatus.OK);
     }
+
     private BCryptPasswordEncoder encoder = new BCryptPasswordEncoder(12);
     @Override
-    public void addUser(User user) {
+    public ResponseEntity<String> addUser(User user) {
+        logger.info("inside addUserMethod");
         User userOne = userRepository.findByEmail(user.getEmail());
         if(userOne != null){
             throw new EmailAlreadyFoundException("Email Already found in database");
         }
         user.setPassword(encoder.encode(user.getPassword()));
         userRepository.save(user);
-        System.out.println("saved to db");
+//        sendNotification();
+        return new ResponseEntity<>("User Added",HttpStatus.OK);
+    }
+
+    @Override
+    public ResponseEntity<String> deleteUserByEmail(String email) {
+        userRepository.deleteByEmail(email);
+        return new ResponseEntity<>("User Deleted", HttpStatus.OK);
     }
 
     @Override
     public UserResponse getUsers(int pageNo, int pageSize, String pageSortBy, String sortDir) {
-
+        logger.info("Inside getUsers method");
         Sort sort = sortDir.equalsIgnoreCase(Sort.Direction.ASC.name()) ? Sort.by(pageSortBy).ascending():
                 Sort.by(pageSortBy).descending();
 
@@ -152,10 +188,10 @@ public class UserServiceImplementation implements UserService {
     // create Feign client and implement
 
 //    @CircuitBreaker(name= "${spring.application.name}" , fallbackMethod = "getDefaultOrders")
-    @Retry(name= "${spring.application.name}" , fallbackMethod = "getDefaultOrders")
+//    @Retry(name= "${spring.application.name}" , fallbackMethod = "getDefaultOrders")
     @Override
     public UserDtoOrders getUsersOrders(Long userId) {
-        logger.info("In the GetUsersOrders");
+        logger.info("inside getUsersOrders method");
         User user = userRepository.findById(userId).orElseThrow(()->new UserNotFoundException("No User Found with this userId: "+userId));
 
 //        ResponseEntity<List<OrderDTO>> responseEntity = restTemplate.exchange(
@@ -173,7 +209,6 @@ public class UserServiceImplementation implements UserService {
 //                .block();
 
         List<OrderDTO> orders = feignAPIClient.getOrdersByUserId(userId);
-        logger.info("Inside getOrderByUserId");
 
 //        orders.forEach( order -> {
 //            List<ProductDTO> products = webClient.get()
@@ -203,26 +238,26 @@ public class UserServiceImplementation implements UserService {
 
     @Override
     @CachePut(value = "getUserById", key = "#id")
-    public UserDTO updateUserById(Long userId, User user) {
-        logger.info("entered into update user by id method");
-        User u = userRepository.findById(userId).orElseThrow(()->new UserNotFoundException("No User Found with this userId: "+userId));
-
-         return AutoUserMapper.MAPPER.mapToDTO(user);
+    public ResponseEntity<UserDTO> updateUserById(Long userId, User user) {
+        logger.info("inside updateUserById method");
+        User user1 = userRepository.findById(userId).orElseThrow(()->new UserNotFoundException("No User Found with this userId: "+userId));
+        return new ResponseEntity<>(AutoUserMapper.MAPPER.mapToDTO(user1), HttpStatus.OK);
     }
 
-//    @Override
-//    public String verify(User user) {
-//        Authentication authentication = authManager.authenticate(new UsernamePasswordAuthenticationToken(user.getEmail(), user.getPassword()));
-//        if (authentication.isAuthenticated()) {
-//            return jwtService.generateToken(user.getEmail());
-//        } else {
-//            return "fail";
-//        }
-//    };
+    @Override
+    public String verify(User user) {
+        logger.info("inside verify method");
+        Authentication authentication = authManager.authenticate(new UsernamePasswordAuthenticationToken(user.getEmail(), user.getPassword()));
+        if (authentication.isAuthenticated()) {
+            return jwtService.generateToken(user.getEmail());
+        } else {
+            return "Verification Failed";
+        }
+    };
 
     public UserDtoOrders getDefaultOrders(String userId, Exception e) {
 
-        logger.info("In the GetDefaultOrders");
+        logger.info("inside getDefaultOrders method");
         User user = userRepository.findById(Long.parseLong(userId)).orElseThrow(()->new UserNotFoundException("No User Found with this userId: "+userId));
 
 
@@ -245,8 +280,8 @@ public class UserServiceImplementation implements UserService {
     // Fetch rollout percentage for the feature
     @Override
     public boolean isFeatureEnabled(Long userId, Long weblabId) {
+        logger.info("inside isFeatureEnabled method");
         WeblabDTO feature = feignWeblabClient.getWeblabById(weblabId);
-        System.out.println("WEBLAB PERCENTAGE"+feature.getWeblabPercentage());
         if ( feature.getWeblabId() > 0 ) {
             long rolloutPercentage = feature.getWeblabPercentage();
             return assignUserToNewFeature(userId, rolloutPercentage);
@@ -256,8 +291,8 @@ public class UserServiceImplementation implements UserService {
 
 
     public boolean assignUserToNewFeature(Long userId, Long rolloutPercentage) {
+        logger.info("inside assignUserToNewFeature method");
         int userHash = Math.abs(userId.hashCode()) % 100;
-        System.out.println("USER HASH VALVE "+userHash);
         return userHash <= rolloutPercentage;
     }
